Add custom favicon
Light bulb icon was chosen as it relates to memory.  Used 32x32 icon
file
to match the original HTML5 Boilerplate version (for HiDPI support).  A
16x16
version should not be necessary.  Removed the default apple-touch-icon
files,
as it's unlikely they will be used.  Credited the icon's source in the
README.  Removed a related comment in index.html.